requirements.txt

fastapi==0.112.2
uvicorn[standard]==0.30.6
psycopg2-binary==2.9.9
redis==5.0.8
prometheus-client==0.20.0
pydantic==2.8.2
python-jose[cryptography]==3.3.0
cryptography==43.0.1
ortools==9.10.4067
